[FEAT] Ability Disable Custom Lines
Feature: Implemented a function that checks either header-line's or
status-line's total number of components so that if there is not any,
Yatline's custom either header-line or status-line will be disabled and
the default Yazi representative will be used. Moreover, String based
component creator function's mode parameter removed because it had never
been used.